Move controls drawer to cladeSelection
